перенес все настройки в page-controller
