+++ src/components/films-wrapper.js
@@ -1,10 +1,6 @@
 import {AbstractComponent} from '../components/abstract-component.js';
 
 export class FilmsSection extends AbstractComponent {
-  constructor() {
-    super();
-  }
-
   getTemplate() {
     return `<section class="films"></div>`;
   }
