+++ src/components/site-menu.js
@@ -1,10 +1,6 @@
 import {AbstractComponent} from "./abstract-component";
 
 export class MenuWrapper extends AbstractComponent {
-  constructor() {
-    super();
-  }
-
   getTemplate() {
     return `<nav class="main-navigation"></nav>`;
   }
