+++ src/components/sort.js
@@ -1,10 +1,6 @@
 import {AbstractComponent} from "./abstract-component";
 
 export class Sort extends AbstractComponent {
-  constructor() {
-    super();
-  }
-
   getTemplate() {
     return `<ul class="sort">
       <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
